Using a u8 to index a 0x8000 array it not a good idea :) thanks to riccardom for reporting.

+++ desmume/src/FIFO.h
@@ -36,8 +36,8 @@ typedef struct
 	BOOL	full;
 	u8		irq;
 
-	u8		sendPos;
-	u8		recvPos;
+	u16		sendPos;
+	u16		recvPos;
 
 	u32		buf[0x8000];
 } FIFO;
